Move document components down to documents dir
